breaking: Remove TargetRPC & use ClientRPC option instead
