+++ Assets/Mirror/Editor/Weaver/Processors/TargetRpcProcessor.cs
@@ -8,70 +8,6 @@ namespace Mirror.Weaver
     /// </summary>
     public static class TargetRpcProcessor
     {
-        // helper functions to check if the method has a NetworkConnection parameter
-        public static bool HasNetworkConnectionParameter(MethodDefinition md)
-        {
-            return md.Parameters.Count > 0 &&
-                   md.Parameters[0].ParameterType.FullName == Weaver.INetworkConnectionType.FullName;
-        }
-
-        /// <summary>
-        /// Generates a skeleton for a TargetRPC
-        /// </summary>
-        /// <param name="td"></param>
-        /// <param name="method"></param>
-        /// <param name="cmdCallFunc"></param>
-        /// <returns>The newly created skeleton method</returns>
-        /// <remarks>
-        /// Generates code like this:
-        /// <code>
-        /// protected static void Skeleton_Test(NetworkBehaviour obj, NetworkReader reader, NetworkConnection senderConnection)
-        /// {
-        ///     if (!obj.netIdentity.server.active)
-        ///     {
-        ///         return;
-        ///     }
-        ///     ((ShipControl) obj).UserCode_Test(reader.ReadSingle(), (int) reader.ReadPackedUInt32());
-        /// }
-        /// </code>
-        /// </remarks>
-        public static MethodDefinition GenerateSkeleton(MethodDefinition md, MethodDefinition userCodeFunc)
-        {
-            var rpc = new MethodDefinition(MethodProcessor.SkeletonPrefix + md.Name, MethodAttributes.Family |
-                    MethodAttributes.Static |
-                    MethodAttributes.HideBySig,
-                    Weaver.voidType);
-
-            ILProcessor worker = rpc.Body.GetILProcessor();
-            Instruction label = worker.Create(OpCodes.Nop);
-
-            NetworkBehaviourProcessor.WriteClientActiveCheck(worker, md.Name, label, "TargetRPC");
-
-            // setup for reader
-            worker.Append(worker.Create(OpCodes.Ldarg_0));
-            worker.Append(worker.Create(OpCodes.Castclass, md.DeclaringType));
-
-            // NetworkConnection parameter is optional
-            if (HasNetworkConnectionParameter(md))
-            {
-                //client.connection
-                worker.Append(worker.Create(OpCodes.Ldarg_0));
-                worker.Append(worker.Create(OpCodes.Call, Weaver.BehaviorConnectionToServerReference));
-            }
-
-            // process reader parameters and skip first one if first one is NetworkConnection
-            if (!NetworkBehaviourProcessor.ReadArguments(md, worker, RemoteCallType.TargetRpc))
-                return null;
-
-            // invoke actual ServerRpc function
-            worker.Append(worker.Create(OpCodes.Callvirt, userCodeFunc));
-            worker.Append(worker.Create(OpCodes.Ret));
-
-            NetworkBehaviourProcessor.AddInvokeParameters(rpc.Parameters);
-            md.DeclaringType.Methods.Add(rpc);
-            return rpc;
-        }
-
         /// <summary>
         /// Replaces the user code with a stub.
         /// Moves the original code to a new method
@@ -111,73 +47,5 @@ namespace Mirror.Weaver
         /// }
         /// </code>
         /// </remarks>
-        public static MethodDefinition GenerateStub(MethodDefinition md, CustomAttribute targetRpcAttr)
-        {
-            MethodDefinition rpc = MethodProcessor.SubstituteMethod(md);
-
-            ILProcessor worker = md.Body.GetILProcessor();
-
-            NetworkBehaviourProcessor.WriteSetupLocals(worker);
-
-            NetworkBehaviourProcessor.WriteCreateWriter(worker);
-
-            // write all the arguments that the user passed to the TargetRpc call
-            // (skip first one if first one is NetworkConnection)
-            if (!NetworkBehaviourProcessor.WriteArguments(worker, md, RemoteCallType.TargetRpc))
-                return null;
-
-            string rpcName = md.Name;
-
-            // invoke SendInternal and return
-            // this
-            worker.Append(worker.Create(OpCodes.Ldarg_0));
-            if (HasNetworkConnectionParameter(md))
-            {
-                // connection
-                worker.Append(worker.Create(OpCodes.Ldarg_1));
-            }
-            else
-            {
-                // null
-                worker.Append(worker.Create(OpCodes.Ldnull));
-            }
-            worker.Append(worker.Create(OpCodes.Ldtoken, md.DeclaringType));
-            // invokerClass
-            worker.Append(worker.Create(OpCodes.Call, Weaver.getTypeFromHandleReference));
-            worker.Append(worker.Create(OpCodes.Ldstr, rpcName));
-            // writer
-            worker.Append(worker.Create(OpCodes.Ldloc_0));
-            worker.Append(worker.Create(OpCodes.Ldc_I4, targetRpcAttr.GetField("channel", 0)));
-            worker.Append(worker.Create(OpCodes.Callvirt, Weaver.sendTargetRpcInternal));
-
-            NetworkBehaviourProcessor.WriteRecycleWriter(worker);
-
-            worker.Append(worker.Create(OpCodes.Ret));
-
-            return rpc;
-        }
-
-        public static bool Validate(MethodDefinition md)
-        {
-            if (md.IsAbstract)
-            {
-                Weaver.Error("Abstract TargetRpc are currently not supported, use virtual method instead", md);
-                return false;
-            }
-
-            if (md.IsStatic)
-            {
-                Weaver.Error($"{md.Name} must not be static", md);
-                return false;
-            }
-
-            if (!NetworkBehaviourProcessor.ProcessMethodsValidateFunction(md))
-            {
-                return false;
-            }
-
-            // validate
-            return NetworkBehaviourProcessor.ProcessMethodsValidateParameters(md, RemoteCallType.TargetRpc);
-        }
     }
 }
