+++ Assets/Mirror/Tests/Editor/Weaver/WeaverMonoBehaviourTests.cs
@@ -1,4 +1,4 @@
-﻿using NUnit.Framework;
+using NUnit.Framework;
 
 namespace Mirror.Weaver.Tests
 {
