--- Assets/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 63ff6fb9db0834ce3999ae706a4d41a7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
